updated from internal repo on 2024/11/06, rev 3d88ac71b14c53f13993848988e2531e13355879

+++ prog/engine/drv/drv3d_DX12/device_context.cpp
@@ -4097,7 +4097,8 @@ void DeviceContext::ExecutionContext::buildTopAccelerationStructure(uint32_t bat
   contextState.resourceStates.flushPendingUAVActions(contextState.graphicsCommandListBarrierBatch, device.currentEventPath().data(),
     device.validatesUserBarriers());
 
-  contextState.resourceActivationTracker.flushAll(contextState.graphicsCommandListBarrierBatch);
+  if (batch_index == 0)
+    contextState.resourceActivationTracker.flushAll(contextState.graphicsCommandListBarrierBatch);
 
   contextState.graphicsCommandListBarrierBatch.execute(contextState.cmdBuffer);
